+++ CHANGELOG.md
@@ -2,11 +2,41 @@
 ===========================
 
 
+## [2.1.0](https://github.com/buession/buession-springcloud/releases/tag/v2.1.0) (2022-08-07)
+
+### 🔨依赖升级
+
+- [依赖库版本升级和安全漏洞修复](https://github.com/buession/buession-parent/releases/tag/v2.1.0)
+- [openfeign] 版本升级至 11.9.1
+
+
+### ⭐ 新特性
+
+- **buession-springcloud-bus：** 新增 buession-springcloud-bus module
+- **buession-springcloud-feign：** 可指定允许转发和忽略的请求头
+
+
+### 🔔 变化
+
+- **buession-springcloud-feign：** Servlet 模式请求头拦截器，ServletClientHeadersRequestInterceptor 将会转发重复请求头的所有值
+
+
+### 🐞 Bug 修复
+
+- **buession-springcloud-feign：** 修复 ReactiveFeignInterceptorConfiguration 自动配置设置成 Type.SERVLET 的 BUG
+
+
+---
+
+
 ## [2.0.2](https://github.com/buession/buession-springcloud/releases/tag/v2.0.2) (2022-07-28)
 
 ### 🔨依赖升级
 
 - [依赖库版本升级和安全漏洞修复](https://github.com/buession/buession-parent/releases/tag/v2.0.2)
+
+
+---
 
 
 ## [2.0.1](https://github.com/buession/buession-springcloud/releases/tag/v2.0.1) (2022-07-18)
@@ -16,6 +46,9 @@
 - [依赖库版本升级和安全漏洞修复](https://github.com/buession/buession-parent/releases/tag/v2.0.1)
 - spring-cloud-starter-alibaba-nacos-config 版本升级至 2.2.8.RELEASE
 - spring-cloud-starter-alibaba-nacos-discovery 版本升级至 2.2.8.RELEASE
+
+
+---
 
 
 ## [2.0.0](https://github.com/buession/buession-springcloud/releases/tag/v2.0.0) (2022-07-08)